+++ src/models/index.ts
@@ -1,7 +1,7 @@
 export * from './Post';
 export * from './FallbackProps';
-export * from './PostsResponse';
-export * from './Query';
+export * from './ApiResponse';
+export * from './QueryParams';
 export * from './User';
 export * from './Comment';
 export * from './FullPost';
